service worker fix try 1
