+++ apps/front-end/public/manifest.json
@@ -1,6 +1,10 @@
 {
-  "short_name": "EG - Project Pragati",
   "name": "EG - Project Pragati",
+  "short_name": "EG - Project Pragati",
+  "start_url": "/",
+  "display": "standalone",
+  "theme_color": "#000000",
+  "background_color": "#ffffff",
   "icons": [
     {
       "src": "images/logos/favicon.ico",
@@ -23,9 +27,5 @@
       "type": "image/png",
       "sizes": "512x512"
     }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
+  ]
 }
