+++ apps/front-end/src/pages/front-end/Dashboard.js
@@ -405,7 +405,9 @@ export default function Dashboard({ userTokenInfo, footerLinks }) {
                 resizeMode="contain"
               />
               <FrontEndTypo.H1 color="textMaroonColor.400" pl="1">
-                Hello World New Changes
+                {process.env.PUBLIC_URL}
+                <br />
+                Hello World New Changes version 2
                 <br />
                 {t("WELCOME")} {facilitator?.first_name},
               </FrontEndTypo.H1>
